requirements.txt
# Google Ads Automation Requirements
# Empire Amplify - Melbourne Digital Marketing Agency

# Core Google Ads SDK
google-ads>=25.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Google API clients
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Environment management
python-dotenv>=1.0.0

# Reporting
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Scheduling
schedule>=1.2.0

# Logging
loguru>=0.7.0
